refactor search result types
